Refactor package, add example
