+++ magefile.go
@@ -5,6 +5,7 @@ package main
 
 import (
 	"github.com/taxibeat/bake/targets/lint/docker"
+	"github.com/taxibeat/bake/targets/prometheus"
 	"github.com/taxibeat/bake/targets/test"
 
 	// mage:import
@@ -20,10 +21,13 @@ import (
 	// mage:import
 	_ "github.com/taxibeat/bake/targets/lint/golang"
 	// mage:import
+	_ "github.com/taxibeat/bake/targets/prometheus"
+	// mage:import
 	_ "github.com/taxibeat/bake/targets/ci"
 )
 
 func init() {
 	docker.DockerFiles = []string{"./Dockerfile"}
+	prometheus.AlertsDir = "./targets/prometheus/examples"
 	test.CoverExcludePatterns = []string{"doc/", "docker/component/testservice/"}
 }
